Create README.md
windows-mobile-hotspot
Automatically start and restart the Windows "Mobile Hotspot" feature on boot as well as in a set interval.

reason
My Windows "Mobile Hotspot" kept on 'breaking' after about 20-30 minutes leading to a lot of applications on my other devices to not load even though they were still connected to the hotspot. I couldn't figure out how to fix it, but restarting the mobile hotspot worked, so I just wanted to automate it.

This shouldn't really be necessary but I couldn't figure out how to get it working without breaking after some time. If you just want the "Mobile Hotspot" to start when your computer launches, perhaps check out the free "NoWiFi" app on the Microsoft store.

other stuff
I would suggest using this with pm2 and pm2-windows-startup.

The restarting of the hotspot itself I did with the help of this Stackoverflow topic.

If you're using a laptop and you're having issues with the hotspot when closing the lid and want it to keep running, make sure you go to "Device Manager" and click "View" and set "Show Hidden Devices" to true. Under "Network Adapters" open all of them and for the ones that have a "Power Management" tab, set "Allow the computer to turn off this device to save power" to false. I also didn't have a battery/power problem, so I set my laptop to never sleep or turn off the display in the "Power Options" found in the "Control Panel" under "Hardware and Sound".